full reference support; Schema.to_dbml()
