--- dlt/helpers/visualize/__init__.py
@@ -1,3 +1,0 @@
-from dlt.helpers.visualize._dbml import schema_to_dbml, export_to_dbml
-
-__all__ = ("schema_to_dbml", "export_to_dbml")
